Split out isTermsEqual
