+++ client/state/posts/utils/is-terms-equal.js
@@ -1,0 +1,22 @@
+/**
+ * External dependencies
+ */
+import { isPlainObject, map, toArray, every, xor } from 'lodash';
+
+/**
+ * Returns truthy if local terms object is the same as the API response
+ *
+ * @param  {object}  localTermEdits local state of term edits
+ * @param  {object}  savedTerms     term object returned from API POST
+ * @returns {boolean}                are there differences in local edits vs saved terms
+ */
+export function isTermsEqual( localTermEdits, savedTerms ) {
+	return every( localTermEdits, ( terms, taxonomy ) => {
+		const termsArray = toArray( terms );
+		const isHierarchical = isPlainObject( termsArray[ 0 ] );
+		const normalizedEditedTerms = isHierarchical ? map( termsArray, 'ID' ) : termsArray;
+		const normalizedKey = isHierarchical ? 'ID' : 'name';
+		const normalizedSavedTerms = map( savedTerms[ taxonomy ], normalizedKey );
+		return ! xor( normalizedEditedTerms, normalizedSavedTerms ).length;
+	} );
+}
